LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/worker-src-self/worker-import.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=617ec011-6a41-42fa-9c97-385700e331dd&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=9198179f-0cb0-4a6b-bf4a-0e40ff6553ff&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=451f7693-847e-4883-956b-5e21db19da2e&path=/mixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/worker.py?action=purge&key=d93322c5-c1cc-4cdc-9c17-ac604108c0cd&path=%2Fmixed-content

Harness Error (TIMEOUT), message = null

PASS Content Security Policy: Expects allowed for worker-import to same-http origin and keep-origin redirection from http context.
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and keep-origin redirection from http context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and no-redirect redirection from http context.
PASS Content Security Policy: Expects allowed for worker-import to same-http origin and no-redirect redirection from http context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and keep-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and no-redirect redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for worker-import to cross-http origin and swap-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for worker-import to cross-http origin and swap-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
TIMEOUT Content Security Policy: Expects blocked for worker-import to same-http origin and swap-origin redirection from http context. Test timed out
NOTRUN Content Security Policy: Expects blocked for worker-import to same-http origin and swap-origin redirection from http context.: securitypolicyviolation
